refactor "note.ts" to be (unit)testable at all
